+++ forui/example/lib/main.dart
@@ -34,7 +34,7 @@ class _ApplicationState extends State<Application> with SingleTickerProviderStat
     locale: const Locale('en', 'US'),
     localizationsDelegates: FLocalizations.localizationsDelegates,
     supportedLocales: FLocalizations.supportedLocales,
-    theme: FThemes.zinc.light.toMaterialTheme(),
+    theme: FThemes.zinc.light.toApproximateMaterialTheme(),
     builder: (context, child) => FTheme(data: FThemes.zinc.light, child: child!),
     home: Scaffold(
       appBar: AppBar(title: const Text('Example')),