+++ forui/CHANGELOG.md
@@ -9,7 +9,7 @@
 * Add `FPickerWheelMixin`.
 * Add `FTimePicker`.
 * Add `FTimePickerController`.
-* Add `FThemeData.toMaterialTheme()`
+* Add `FThemeData.toApproximateMaterialTheme()`
 
 ### Changes